+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # JEC GlassCat Metadata Project: Update Release Notes
 
+<a name="jec-glasscat-metadata-1.0.2"></a>
+## **1.0.2** (2018-03-28)
+
+### Bug Fixes
+
+### Features
+
+- Replacing the string representation of the creation date by a timestamp
+
 <a name="jec-glasscat-metadata-1.0.1"></a>
 ## **1.0.1** (2018-03-25)